requirements.txt
backports.tarfile==1.2.0
brotli==1.1.0
chardet==5.2.0
defusedxml==0.7.1
importlib-metadata==8.0.0
jaraco.collections==5.1.0
jinja2==3.1.5
packaging==24.2
pandas==2.2.3
platformdirs==4.2.2
pyyaml==6.0.2
tomli==2.0.1
weasyprint==64.0
zopfli==0.2.3.post1